debugging no image errors

+++ app/Http/Controllers/admin/BooksController.php
@@ -7,6 +7,7 @@ use App\Models\books;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
+
 class BooksController extends Controller
 {
     /**
@@ -15,7 +16,7 @@ class BooksController extends Controller
     public function index()
     {
         $books = books::all();
-        return view("admin.book.index",compact("books"));
+        return view("admin.book.index", compact("books"));
     }
 
     /**
@@ -37,19 +38,19 @@ class BooksController extends Controller
                 "title" => "required|unique:books,title",
                 "semester_id" => "required",
                 "category_id" => "required",
-                "document"=> "required"
+                "document" => "required"
             ],
             [
                 "title.required" => "Title is required field",
                 "semester_id.required" => "Select Semester",
                 "category_id.required" => "Select Category",
-                "unique"=> "book Already exists.",
-                "document.required"=> "Document is required"
-        ],
+                "unique" => "book Already exists.",
+                "document.required" => "Document is required"
+            ],
         );
         $book = new books();
         $book->title = Str::title($request->title);
-        $book->slug = Str::slug($request->title,"-");
+        $book->slug = Str::slug($request->title, "-");
         $book->semester_id = $request->semester_id;
         $book->category_id = $request->category_id;
         if ($request->hasfile("photo")) {
@@ -57,6 +58,8 @@ class BooksController extends Controller
             $newfile = time() . "." . $file->GetClientOriginalExtension();
             $file->move("images/book", $newfile);
             $book->photo = ("images/book/$newfile");
+        } else {
+            $book->photo = ("images/no-image.png");
         }
         if ($request->hasfile("document")) {
             $doc = $request->document;
@@ -81,8 +84,8 @@ class BooksController extends Controller
      */
     public function edit(string $id)
     {
-          $book = books::find($id);
-        return view("admin.book.edit",compact("book"));
+        $book = books::find($id);
+        return view("admin.book.edit", compact("book"));
     }
 
     /**
@@ -90,7 +93,7 @@ class BooksController extends Controller
      */
     public function update(Request $request, string $id)
     {
-         $request->validate(
+        $request->validate(
             [
                 "title" => "required",
                 "semester_id" => "required",
@@ -98,37 +101,39 @@ class BooksController extends Controller
             [
                 "title.required" => "Title is required field",
                 "semester_id.required" => "Select Semester",
-        ],
+            ],
         );
-        try{$book = books::find($id);
-        $book->title = Str::title($request->title);
-        $book->slug = Str::slug($request->title,"-");
-        $book->semester_id = $request->semester_id;
-        $book->category_id = $request->category_id;
-        if ($request->hasfile("document")) {
-            $doc = $request->document;
-            $olddoc =  $book->document;
-            if (File::exists(public_path($olddoc))) {
-                File::delete(public_path($olddoc));
+        try {
+            $book = books::find($id);
+            $book->title = Str::title($request->title);
+            $book->slug = Str::slug($request->title, "-");
+            $book->semester_id = $request->semester_id;
+            $book->category_id = $request->category_id;
+            if ($request->hasfile("document")) {
+                $doc = $request->document;
+                $olddoc =  $book->document;
+                if (File::exists(public_path($olddoc))) {
+                    File::delete(public_path($olddoc));
+                }
+                $newdoc = time() . "." . $doc->GetClientOriginalExtension();
+                $doc->move("doc/book", $newdoc);
+                $book->document = ("doc/book/$newdoc");
             }
-            $newdoc = time() . "." . $doc->GetClientOriginalExtension();
-            $doc->move("doc/book", $newdoc);
-            $book->document = ("doc/book/$newdoc");
-        }
-        $book->update();
-    }catch(\illuminate\database\QueryException $e){
-        $errorcode = $e->errorInfo[1];
-        if($errorcode==1062){
-             toast('Operation failed. book Already exists!', 'warning')->position('bottom-end');
-            return back();
-        }
-
-    };
-    if ($request->hasfile("photo")) {
+            $book->update();
+        } catch (\illuminate\database\QueryException $e) {
+            $errorcode = $e->errorInfo[1];
+            if ($errorcode == 1062) {
+                toast('Operation failed. book Already exists!', 'warning')->position('bottom-end');
+                return back();
+            }
+        };
+        if ($request->hasfile("photo")) {
             $file = $request->photo;
             $oldfile =  $book->photo;
-            if (File::exists(public_path($oldfile))) {
-                File::delete(public_path($oldfile));
+            if ($oldfile != 'images/no-image.png') {
+                if (File::exists(public_path($oldfile))) {
+                    File::delete(public_path($oldfile));
+                }
             }
             $newfile = time() . "." . $file->GetClientOriginalExtension();
             $file->move("images/book", $newfile);
@@ -143,11 +148,13 @@ class BooksController extends Controller
      */
     public function destroy(string $id)
     {
-        $book= books::find($id);
-         $oldfile =  $book->photo;
-         $olddoc =  $book->document;
-        if (File::exists(public_path($oldfile))) {
-            File::delete(public_path($oldfile));
+        $book = books::find($id);
+        $oldfile =  $book->photo;
+        $olddoc =  $book->document;
+        if ($oldfile != 'images/no-image.png') {
+            if (File::exists(public_path($oldfile))) {
+                File::delete(public_path($oldfile));
+            }
         }
         if (File::exists(public_path($olddoc))) {
             File::delete(public_path($olddoc));
